+++ common/src/main/java/me/pandamods/fallingtrees/FallingTrees.java
@@ -3,7 +3,6 @@ package me.pandamods.fallingtrees;
 import me.pandamods.fallingtrees.config.ClientConfig;
 import me.pandamods.fallingtrees.config.CommonConfig;
 import me.pandamods.fallingtrees.config.FallingTreesConfig;
-import me.pandamods.fallingtrees.config.ModConfig;
 import me.pandamods.fallingtrees.event.EventHandler;
 import me.pandamods.fallingtrees.network.PacketHandler;
 import me.pandamods.fallingtrees.registry.EntityRegistry;