+++ common/src/main/java/me/pandamods/fallingtrees/config/classes/AlgorithmConfig.java
@@ -1,0 +1,5 @@
+package me.pandamods.fallingtrees.config.classes;
+
+public class AlgorithmConfig {
+
+}