+++ Changelog.md
@@ -1,3 +1,12 @@
 ## Version 0.8 Changelog
+* **What's been recoded**
+  * Config
+
+
+* **What's been changed**
+  * Config file split into 2 files Client and Common
+
+
 * **What's been added**
+  * Algorithm Config options
   * Added Block Filter to config